test: strengthen autogrow persistence assertions
Amp-Thread-ID: https://ampcode.com/threads/T-01a09436-7f44-75c8-8b68-ba9288bbd2de
